refactor(query): extract commands logic into query layer RPC pattern
- Move all command implementations from commands.ts to query/commands.ts
- Convert commands to mutations following query layer patterns
- Move saveRecordingAndTranscribeTransform to processRecordingPipeline
- Move uploadRecording from recordings.ts to commands.ts
- Update commands.ts to use rpc.commands.* calls
- Add commands to RPC namespace in query/index.ts
- Update +page.svelte to use rpc.commands.uploadRecording
- Keep processRecordingPipeline internal, not exposed in RPC
